+++ backend/src/certs/mod.rs
@@ -1,3 +1,4 @@
 pub mod tls_cert;
 pub mod ssh_cert;
 pub mod common;
+pub mod crl;
